Migrate upgrade cluster plan
